Assignment-4/pokemon-trading/requirements.txt
flask #APIs
requests #'Front-end'
mysql-connector-python #SQL engine (db_utils)
